# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Node CRUD API",
    "version": "1.0.0",
    "description": "Authentication and Task management API",
    "contact": { "name": "API Support" }
  },
  "servers": [
    { "url": "http://localhost:5001", "description": "Development server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": { "type": "string", "example": "johndoe" },
          "email": { "type": "string", "example": "john@example.com" },
          "password": { "type": "string", "example": "password123" }
        }
      },
      "Task": {
        "type": "object",
        "required": ["title", "description"],
        "properties": {
          "title": { "type": "string", "example": "First Task" },
          "description": { "type": "string", "example": "This is my first task" },
          "status": { "type": "string", "enum": ["pending", "in-progress", "completed"], "example": "pending" }
        }
      }
    }
  },
  "tags": [
    { "name": "Authentication", "description": "User authentication endpoints" },
    { "name": "Tasks", "description": "Task management endpoints" }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": { "description": "User already exists" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "john@example.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/tasks": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Create a new task",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } }
        },
        "responses": {
          "201": { "description": "Task created successfully" },
          "400": { "description": "Missing title or description" }
        }
      },
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all tasks of the logged-in user",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of tasks" } }
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get task by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" }, "description": "Task ID" }
        ],
        "responses": {
          "200": { "description": "Task details" },
          "400": { "description": "Invalid task ID" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "summary": "Update a task",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" }, "description": "Task ID" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } }
        },
        "responses": {
          "200": { "description": "Task updated successfully" },
          "400": { "description": "Invalid task ID" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete a task",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" }, "description": "Task ID" }
        ],
        "responses": {
          "200": { "description": "Task deleted successfully" },
          "400": { "description": "Invalid task ID" },
          "404": { "description": "Task not found" }
        }
      }
    }
  }
}